+++ h20/lib/presentation/screens/daily_report_screen.dart
@@ -3,12 +3,27 @@ import 'package:flutter/material.dart';
 class DailyReportScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    int totalWaterConsumed = 1200; // in milliliters
+
     return Scaffold(
       appBar: AppBar(
-        title: Text('Daily Report'),
+        title: Text('Daily Report  '),
       ),
       body: Center(
-        child: Text('This is the Daily Report Screen'),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Text(
+              'Total Water Consumed Today ',
+              style: TextStyle(fontSize: 20),
+            ),
+            SizedBox(height: 10),
+            Text(
+              '$totalWaterConsumed ml',
+              style: TextStyle(fontSize: 40, fontWeight: FontWeight.bold),
+            ),
+          ],
+        ),
       ),
     );
   }
