+++ h20/lib/presentation/screens/weekly_report_screen.dart
@@ -8,7 +8,7 @@ class WeeklyReportScreen extends StatelessWidget {
         title: Text('Weekly Report'),
       ),
       body: Center(
-        child: Text('This is the Weekly Report Screen'),
+        child: Text('This is the Weekly  '),
       ),
     );
   }
